requirements.txt
# Core dependencies
streamlit>=1.28.0
numpy>=1.20.0
Pillow>=9.0.0

# OpenVINO dependencies
openvino>=2023.1.0
openvino-genai>=0.1.0

# Additional utilities
python-dotenv>=1.0.0
typing-extensions>=4.0.0
requests>=2.28.0

# Optional - performance monitoring
psutil>=5.9.0

# Optional - image processing enhancements
opencv-python>=4.6.0

pytube